IRSA Fire Safety addFireAccountViolationTable version
This is an edited version of the original Fire Safety IRSA script which will remove violations that are set to Complied in the guidesheet from the Violations ASIT on record. This allows for Inspection Reports to effectively ignore Complied violations. Users will then be expected to navigate into past inspections to see previous violations that had since complied. This script is intended to be used as a mitigation/back up if need be.